backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
boto3
bcrypt
pytest
pytest-cov
httpx
